Adding Tax and delivery fees to UI and Store

+++ Behaviors/NumberValidationBehaviour.cs
@@ -18,9 +18,9 @@ namespace voltaire.Behaviors
 
 		void OnEntryTextChanged(object sender, TextChangedEventArgs args)
 		{
-			int result;
+            double result;
 
-			bool isValid = int.TryParse(args.NewTextValue, out result);
+            bool isValid = double.TryParse(args.NewTextValue, out result);
 
             ((Entry)sender).TextColor = isValid ? (Color) App.Current.Resources["turquoiseBlue"] : Color.Red;
 		}
